1. post id type 변경 -> Long 2. 추천 기능 프론트도 추가 3. 이미 추천 했을 경우 처리

+++ src/main/java/com/inspire12/homepage/controller/community/CommentController.java
@@ -25,7 +25,7 @@ public class CommentController {
 
         commentService.saveByRequest(requestBody);
         ObjectNode responseBody = objectMapper.createObjectNode();
-        List<CommentMsg> comments = commentService.getComments(requestBody.get("article_id").asInt());
+        List<CommentMsg> comments = commentService.getComments(requestBody.get("article_id").asLong());
         responseBody.putPOJO("comments", comments);
         return new ResponseEntity<ObjectNode>(responseBody, HttpStatus.OK);
     }
