+++ src/main/java/com/inspire12/homepage/repository/CommentRepository.java
@@ -15,11 +15,11 @@ public interface CommentRepository extends JpaRepository<Comment, Integer> {
     List<Comment> findAllByArticleId(int article);
 
     @Query(value = "select * from comment where article_id=:articleId order by no desc, grpord asc", nativeQuery = true)
-    List<Comment> selectCommentByArticleOrder(@Param("articleId") int articleId);
+    List<Comment> selectCommentByArticleOrder(@Param("articleId") Long articleId);
     @Modifying
     @Transactional
     @Query(value = "INSERT INTO comment (article_id, username, content, `no`) VALUES (:article_id, :username, :content, (select last_insert_id()+1))", nativeQuery = true)
-    void insertByRequest(@Param("article_id") int articleId, @Param("username") String userId, @Param("content") String content);
+    void insertByRequest(@Param("article_id") Long articleId, @Param("username") String userId, @Param("content") String content);
 
     @Modifying
     @Transactional
